+++ ckan/logic/action/delete.py
@@ -1,9 +1,12 @@
 '''API functions for deleting data from CKAN.'''
 
+import sqlalchemy as sqla
+
 import ckan.logic
 import ckan.logic.action
 import ckan.plugins as plugins
 import ckan.lib.dictization.model_dictize as model_dictize
+from ckan import authz
 
 from ckan.common import _
 
@@ -44,6 +47,9 @@ def user_delete(context, data_dict):
 def package_delete(context, data_dict):
     '''Delete a dataset (package).
 
+    This makes the dataset disappear from all web & API views, apart from the
+    trash.
+
     You must be authorized to delete the dataset.
 
     :param id: the id or name of the dataset to delete
@@ -72,6 +78,46 @@ def package_delete(context, data_dict):
 
     entity.delete()
     model.repo.commit()
+
+
+def dataset_purge(context, data_dict):
+    '''Purge a dataset.
+
+    .. warning:: Purging a dataset cannot be undone!
+
+    Purging a database completely removes the dataset from the CKAN database,
+    whereas deleting a dataset simply marks the dataset as deleted (it will no
+    longer show up in the front-end, but is still in the db).
+
+    You must be authorized to purge the dataset.
+
+    :param id: the name or id of the dataset to be purged
+    :type id: string
+
+    '''
+    model = context['model']
+    id = _get_or_bust(data_dict, 'id')
+
+    pkg = model.Package.get(id)
+    context['package'] = pkg
+    if pkg is None:
+        raise NotFound('Dataset was not found')
+
+    _check_access('dataset_purge', context, data_dict)
+
+    members = model.Session.query(model.Member) \
+                   .filter(model.Member.table_id == pkg.id) \
+                   .filter(model.Member.table_name == 'package')
+    if members.count() > 0:
+        for m in members.all():
+            m.purge()
+
+    pkg = model.Package.get(id)
+    # no new_revision() needed since there are no object_revisions created
+    # during purge
+    pkg.purge()
+    model.repo.commit_and_remove()
+
 
 def resource_delete(context, data_dict):
     '''Delete a resource from a dataset.
@@ -397,12 +443,34 @@ def _group_or_org_purge(context, data_dict, is_org=False):
     else:
         _check_access('group_purge', context, data_dict)
 
-    members = model.Session.query(model.Member)
-    members = members.filter(model.Member.group_id == group.id)
+    if is_org:
+        # Clear the owner_org field
+        datasets = model.Session.query(model.Package) \
+                        .filter_by(owner_org=group.id) \
+                        .filter(model.Package.state != 'deleted') \
+                        .count()
+        if datasets:
+            if not authz.check_config_permission('ckan.auth.create_unowned_dataset'):
+                raise ValidationError('Organization cannot be purged while it '
+                                      'still has datasets')
+            pkg_table = model.package_table
+            # using Core SQLA instead of the ORM should be faster
+            model.Session.execute(
+                pkg_table.update().where(
+                    sqla.and_(pkg_table.c.owner_org == group.id,
+                              pkg_table.c.state != 'deleted')
+                ).values(owner_org=None)
+            )
+
+    # Delete related Memberships
+    members = model.Session.query(model.Member) \
+                   .filter(sqla.or_(model.Member.group_id == group.id,
+                                    model.Member.table_id == group.id))
     if members.count() > 0:
-        model.repo.new_revision()
+        # no need to do new_revision() because Member is not revisioned, nor
+        # does it cascade delete any revisioned objects
         for m in members.all():
-            m.delete()
+            m.purge()
         model.repo.commit_and_remove()
 
     group = model.Group.get(id)
@@ -418,6 +486,8 @@ def group_purge(context, data_dict):
     Purging a group completely removes the group from the CKAN database,
     whereas deleting a group simply marks the group as deleted (it will no
     longer show up in the frontend, but is still in the db).
+
+    Datasets in the organization will remain, just not in the purged group.
 
     You must be authorized to purge the group.
 
@@ -436,6 +506,9 @@ def organization_purge(context, data_dict):
     database, whereas deleting an organization simply marks the organization as
     deleted (it will no longer show up in the frontend, but is still in the
     db).
+
+    Datasets owned by the organization will remain, just not in an
+    organization any more.
 
     You must be authorized to purge the organization.