+++ ckan/pastertemplates/__init__.py
@@ -51,6 +51,11 @@ class CkanextTemplate(Template):
     def check_vars(self, vars, cmd):
         vars = Template.check_vars(self, vars, cmd)
 
+        # workaround for a paster issue https://github.com/ckan/ckan/issues/2636
+        # this is only used from a short-lived paster command
+        reload(sys)
+        sys.setdefaultencoding('utf-8')
+
         if not vars['project'].startswith('ckanext-'):
             print "\nError: Project name must start with 'ckanext-'"
             sys.exit(1)
@@ -63,7 +68,7 @@ class CkanextTemplate(Template):
         keywords = [keyword for keyword in keywords
                     if keyword not in ('ckan', 'CKAN')]
         keywords.insert(0, 'CKAN')
-        vars['keywords'] = ' '.join(keywords)
+        vars['keywords'] = u' '.join(keywords)
 
         # For an extension named ckanext-example we want a plugin class
         # named ExamplePlugin.