+++ ckan/tests/legacy/functional/test_group.py
@@ -13,7 +13,7 @@ class TestGroup(FunctionalTestCase):
 
     @classmethod
     def setup_class(self):
-        search.clear()
+        search.clear_all()
         model.Session.remove()
         CreateTestData.create()
 
@@ -116,7 +116,6 @@ class TestRevisions(FunctionalTestCase):
         self.grp = model.Group(name=self.name)
         self.grp.description = self.description[0]
         model.Session.add(self.grp)
-        model.setup_default_user_roles(self.grp)
         model.repo.commit_and_remove()
 
         # edit pkg